Update gcp_compute_disk.py
This commit solve a problem that whe have been facing when trying to create a disk from a snashot. 

gcp_compute_module allow field sourceSnapshot but it's not sending to gcp api.

Now you can pass selflink or gcp_compute_snapshot object and it will be sended correctly to gcp api.